+++ build_habitjewel.py
@@ -46,13 +46,18 @@ chmod 755 /opt/habitjewel/habitjewel.py
     #    """ #Set here your pre install script
     #  p.preremove="""#!/bin/sh
     #  chmod +x /usr/bin/mclock.py""" #Set here your pre remove script
-    version = "0.7.0"          #Version of your software, for example "1.2.0" or "0.8.2"
+    version = "0.7.1"          #Version of your software, for example "1.2.0" or "0.8.2"
     build = "1"                #Build number, for example "1" for the first build of this version of your software. Increment for later re-builds of the same version of your software.
                                 #Text with changelog information to be displayed in the package "Details" tab of the Maemo Application Manager
-    changeloginformation = """* Implemented countdown timer, accessible for time-based habits on current date only
-* Refactored history gathering
-* Tweaked mini bar graph display
-* Disabled main menu Stats button (for now)
+    changeloginformation = """[0.7.1] - 2016-09-13
+
+### Added
+- Icon on timer start/stop button
+
+### Changed
+- Nicer formatting and colouring of habits in day habits list
+- Other display tweaks
+- Moved change log from main script to CHANGELOG.md
 """
     dir_name = "src"            #Name of the subfolder containing your package source files (for example, usr\share\icons\hicolor\scalable\myappicon.svg, usr\lib\myapp\somelib.py). We suggest to leave it named src in all projects and will refer to that in the wiki article on maemo.org
     #Thanks to DareTheHair from talk.maemo.org for this snippet that recursively builds the file list.