+++ make.py
@@ -4,7 +4,7 @@ import pypackager
 import os
 p=pypackager.PyPackager("habitjewel") #package name
 p.display_name = 'HabitJewel' #package display name in HAM
-p.version = '0.7.0' #package version
+p.version = '0.7.1' #package version
 p.buildversion = 'local1' #package build version
 p.description="""Record and track your progress in achieving your daily habits.""" #package description
 p.author='Ashley Hooper' #package author
@@ -31,10 +31,15 @@ chmod 755 /opt/habitjewel/habitjewel.py
 
 #p.postremove="""#!/bin/sh
 #"""
-p.changelog ="""* Implemented countdown timer, accessible for time-based habits on current date only
-* Refactored history gathering
-* Tweaked mini bar graph display
-* Disabled main menu Stats button (for now)
+p.changelog ="""[0.7.1] - 2016-09-13
+
+### Added
+- Icon on timer start/stop button
+
+### Changed
+- Nicer formatting and colouring of habits in day habits list
+- Other display tweaks
+- Moved change log from main script to CHANGELOG.md
 """ # the changelog
 dir_name='src' #src directory name
 #Here we will loop in all files, directory in src to add it to pkg